LayoutTests/platform/chromium/fast/forms/time/time-stepup-stepdown-from-renderer-expected.txt

Check stepping-up and -down for time input fields from renderer. No cases of empty initial values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Function arguments are (value, step, {min or max}, [stepCount]).
Normal cases
FAIL stepUp("20:13", null, null) should be 20:14. Was 20:13.
FAIL stepDown("20:13", null, null) should be 20:12. Was 20:13.
FAIL stepUp("20:13", null, null, 10) should be 20:23. Was 20:13.
FAIL stepDown("20:13", null, null, 11) should be 20:02. Was 20:13.
FAIL stepUp("20:13", "4", null, 2) should be 20:13:08. Was 20:13.
FAIL stepDown("20:13", "4", null, 3) should be 20:12:48. Was 20:13.
Step=any
FAIL stepUp("20:13", "any", null) should be 20:14. Was 20:13.
FAIL stepDown("20:13", "any", null) should be 20:12. Was 20:13.
Overflow/underflow
FAIL stepUp("20:13", "3.40282346e+38", null) should be 23:59:59.999. Was 20:13.
FAIL stepDown("20:13", "3.40282346e+38", null) should be 00:00:00. Was 20:13.
PASS stepUp("20:13", "1", "20:13") is "20:13"
PASS stepDown("20:13", "1", "20:13") is "20:13"
PASS stepUp("23:59", null, null) is "23:59"
PASS stepDown("00:00", null, null) is "00:00"
stepDown()/stepUp() for stepMismatch values
FAIL stepDown("20:13", "3", "20:12:56") should be 20:12:59. Was 20:13.
FAIL stepUp("00:13", "7", "") should be 00:13:04. Was 00:13.

PASS successfullyParsed is true

TEST COMPLETE
